fix: reparacion de anchores , componente link , add: topbar en diferentes paginas.
