+++ frontend/src/app/admin/readme.md
@@ -1,0 +1,1 @@
+# Rutas Administrativa
