+++ frontend/src/app/dashboard/perfil/components/FooterProfile.tsx
