+++ frontend/src/app/dashboard/readme.md
@@ -1,0 +1,1 @@
+# ruta Dashboard para usuarios autotificados
