--- frontend/src/app/login/page.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import LoginForm from "@/components/froms/LoginForm";
-
-const LoginPage = () => {
-  return (
-    <section className="flex items-center justify-center min-h-screen bg-[url('/img/bg_lr.png')] bg-cover bg-center">
-      <LoginForm />
-    </section>
-  );
-};
-
-export default LoginPage;
