--- frontend/src/app/testvw/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Page() {
-  return (
-    <div>
-      {/* <FooterGeneral /> */}
-    </div>
-  )
-}
-
-export default Page
