+++ frontend/src/components/containers/footer-404.tsx
@@ -3,19 +3,19 @@ import React from 'react'
 import logort from '@/../public/logort.png'
 import Link from 'next/link'
 
-function FooterGeneral() {
+function Foote404() {
   return (
-    <footer className='w-full p-10 bg-primary-variant-2'>
+    <footer className='w-full p-10 bg-primary-variant-3'>
       <div className='flex flex-col'>
         <div className='flex flex-row'>
           <Image src={logort} alt='RT' width={175} height={175} />
           <div className="ml-auto text-white flex flex-row gap-16 md:gap-8 sm:gap-2 sm:text-xs overflow-ellipsis">
             {[
               {href:'/',title:'INICIO',},
-              {href:'/public/nosotros',title:'SOBRE NOSOTROS',},/*PENDIENTE*/
-              {href:'/public/contacto',title:'CONTACTO',}, /*PENDIENTE*/
-              {href:'/public/soporte',title:'SOPORTE',},
-              {href:'/public/terminos',title:'TÉRMINOS Y CONDICIONES',}, /*PENDIENTE*/
+              {href:'/nosotros',title:'SOBRE NOSOTROS',},/*PENDIENTE*/
+              {href:'/contacto',title:'CONTACTO',}, /*PENDIENTE*/
+              {href:'/soporte',title:'SOPORTE',},
+              {href:'/terminos',title:'TÉRMINOS Y CONDICIONES',}, /*PENDIENTE*/
             ].map((item,i) => (
               <Link href={item.href} className='h-2' key={i}>
                 {item.title}
@@ -28,15 +28,15 @@ function FooterGeneral() {
           <h3 className='text-xs text-white'>
             © 2024 Retrueque - Todos los derechos reservados &nbsp;/&nbsp; 
             {/* PENDIENTE */}
-            <Link href='/public/legal' className='mr-2 hover:underline hover:font-bold'>
+            <Link href='/legal' className='mr-2 hover:underline hover:font-bold'>
               Aviso Legal
             </Link> - 
             {/* PENDIENTE */}
-            <Link href='public/equipo' className='mx-2 hover:underline hover:font-bold'>
+            <Link href='/equipo' className='mx-2 hover:underline hover:font-bold'>
               Equipo de Desarrollo
             </Link> - 
             {/* PENDIENTE */}
-            <Link href='public/privacidad' className='mx-2 hover:underline hover:font-bold'>
+            <Link href='/privacidad' className='mx-2 hover:underline hover:font-bold'>
               Política de Privacidad
             </Link>
           </h3>
@@ -46,4 +46,4 @@ function FooterGeneral() {
   )
 }
 
-export default FooterGeneral
+export default Foote404
